Update to game version v193.3693_28985. Decreased workers value for MVP score

+++ com.wsunitstats.exporter/src/main/java/com/wsunitstats/exporter/model/json/gameplay/submodel/CollisionJsonModel.java
@@ -9,7 +9,7 @@ import lombok.ToString;
 @ToString
 public class CollisionJsonModel {
     private Integer elasticity;
-    private Integer tags;
+    private Long tags;
     private Integer radius;
     private Integer weight;
 }
